requirements.txt
streamlit
pandas
numpy
pymysql
sqlalchemy
python-dotenv
requests
opencv-python
ultralytics
scikit-learn
